% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fitbit.R
\name{clean_fitbit}
\alias{clean_fitbit}
\title{Cleans fitbit data}
\usage{
clean_fitbit(fitbit_dat, wear_time, date_of_birth)
}
\arguments{
\item{fitbit_dat}{a data.table representing the activity_summary table containing at least the following columns person_id, date, steps}

\item{wear_time}{a data.table from the output of the aou.phenotyper2::wear_time algorithm}

\item{data_of_birth}{a data.table containing person_id and date_of_birth columns}
}
\description{
Cleans fitbit data
}
\examples{
\dontrun{
library(aou.phenotyper2)
output_folder <- "datasets"
fitbit(output_folder)
demographics(output_folder)
wear_time(output_folder)
fitbit_dat <- read_bucket("datasets/fitbit.csv")
wear_time_dat <- read_bucket("datasets/wear_time.csv")
demographics_dat <- read_bucket("datasets/demographics.csv")
fitbit_cleaned <- clean_fitbit(fitbit_dat,wear_time_dat,demographics[,c("person_id","date_of_birth")])
}
}
