% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fitbit_with_ehr.R
\name{clean_fitbit_with_ehr}
\alias{clean_fitbit_with_ehr}
\title{Query and clean fitbit data and combine with ehr}
\usage{
clean_fitbit_with_ehr(output_folder = "datasets", bucket = NULL)
}
\arguments{
\item{output_folder}{the folder to check for cached queries}

\item{bucket}{the bucket to check for cached queries}
}
\description{
Query and clean fitbit data and combine with ehr
}
\examples{
\dontrun{
fitbit_cleaned_combined <- clean_fitbit_with_ehr()
}
}
