% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sleep.R
\name{clean_sleep}
\alias{clean_sleep}
\title{Cleans sleep data}
\usage{
clean_sleep(sleep_dat, date_of_birth)
}
\arguments{
\item{sleep_dat}{a data.table representing the sleep_summary table containing at least the following columns person_id, is_main_sleep, minute_asleep}

\item{data_of_birth}{a data.table containing person_id and date_of_birth columns}
}
\description{
Cleans sleep data
}
\examples{
\dontrun{
library(aou.phenotyper2)
output_folder <- "datasets"
sleep(output_folder)
demographics(output_folder)
fitbit_dat <- read_bucket("datasets/fitbit.csv")
demographics_dat <- read_bucket("datasets/demographics.csv")
sleep_cleaned <- clean_fitbit(fitbit_dat,demographics[,c("person_id","date_of_birth")])
}
}
