% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fitbit_cox.R
\name{format_fitbit_cox}
\alias{format_fitbit_cox}
\title{Format Fitbit data for time varying Cox model}
\usage{
format_fitbit_cox(fitbit, dx, last_medical_encounter)
}
\arguments{
\item{fitbit}{data.table with columns: person_id, date, steps, sedentary_minutes, lightly_active_minutes, fairly_active_minutes, very_active_minutes}

\item{dx}{data.table with columns: person_id, dx_entry_date, dx_status}

\item{last_medical_encounter}{data.table with columns: person_id, last_medical_encounter_entry_date}
}
\value{
A data.table
}
\description{
Format Fitbit data for time varying Cox model
}
