% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fitbit_sleep_activity_cox.R
\name{format_fitbit_sleep_activity_cox}
\alias{format_fitbit_sleep_activity_cox}
\title{Format Fitbit activity and sleep data for time varying Cox model}
\usage{
format_fitbit_sleep_activity_cox(
  sleep_pa,
  dx,
  last_medical_encounter,
  cull_list,
  cull_baseline = TRUE
)
}
\arguments{
\item{sleep_pa}{data.table with columns: person_id, date, minute_asleep, minute_rem, minute_deep, minute_light, minute_in_bed, steps}

\item{dx}{data.table with columns: person_id, dx_entry_date, dx_status}

\item{last_medical_encounter}{data.table with columns: person_id, last_medical_encounter_entry_date}

\item{cull_list}{a character vector containing variables to perform exclusions on with months less than 15 days of observations.}

\item{cull_baseline}{should there be at least 15 observations in the baseline monitoring period?}
}
\value{
A data.table
}
\description{
Format Fitbit activity and sleep data for time varying Cox model
}
