% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fitbit_sleep_cox.R
\name{format_sleep_cox}
\alias{format_sleep_cox}
\title{Format Fitbit sleep data for time varying Cox model}
\usage{
format_sleep_cox(sleep, dx, last_medical_encounter)
}
\arguments{
\item{dx}{data.table with columns: person_id, dx_entry_date, dx_status}

\item{last_medical_encounter}{data.table with columns: person_id, last_medical_encounter_entry_date}

\item{fitbit}{data.table with columns: person_id, date, minute_asleep, minute_rem, minute_deep}
}
\value{
A data.table
}
\description{
Format Fitbit sleep data for time varying Cox model
}
