% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_cohort_iterative.R
\name{match_cohort_iterative}
\alias{match_cohort_iterative}
\title{Matches cases with control using activity data iteratively}
\usage{
match_cohort_iterative(
  cases_dat,
  match_rate,
  controls_dat,
  window_range = NULL,
  strict_cols = NULL,
  flex_cols = NULL,
  range_cols = NULL,
  baseline_cols = NULL,
  fixed_baselines = NULL,
  baseline_range = NULL,
  anchor_col_name = NULL,
  min_days_in_baseline = NULL,
  rank = NULL
)
}
\arguments{
\item{match_rate}{the number of control to match each case}

\item{window_range}{the window range to require available activity data around the anchor date}

\item{strict_cols}{vector of column names to match control exactly such as sex or race}

\item{flex_cols}{vector of column names to match based on similarity}

\item{range_cols}{named list of column names to match strictly in a range such as age}

\item{baseline_cols}{vector of column names to match based on similarity of a mean in baseline_range parameter}

\item{fixed_baselines}{vector of column names to match based on similarity closest to start of baseline_range}

\item{baseline_range}{baseline_range given in time to event to compute baseline averages and find closest fixed baselines}

\item{anchor_col_name}{the column name for the anchor date used for baselines and windowing}

\item{min_days_in_baseline}{the required number of days in baseline applied to both cases and control}

\item{rank}{ordering column for cases before the iteration}

\item{cohort_dat}{a data.table representing the activity_summary table for the cohort to match containing at least the following columns person_id, date}

\item{control_dat}{a data.table representing the activity_summary table for the available control candidates to match containing at least the following columns person_id, date}
}
\value{
a dataframe of matched control with corresponding cases and differences between flexible matched columns and baselines
}
\description{
Matches cases with control using activity data iteratively
}
\details{
This algorithm follows the following steps to match the given cases to a control cohort:
First it subsets the cases using window_range and min_days_in_baseline
Then orders cases using rank parameter (if given) and iteratively matches each case to match_rate many control cases following the following steps:
First it subsets the controls using strict_cols and range_cols
Computes baselines and fixed_baselines for controls. The event date for controls is set to event date of case that is being matched at the iterative step
Finds the best match_rate many candidates using matchit library  using flex_cols, baseline_cols, and fixed_baselines
Returns controls for each cases in one datatable
If there is not enough control cases after candidate selection using strict_cols and range_cols then all controls are chosen and if none available then that case won't appear in resulting datatable
}
\examples{
\dontrun{
control_long_covid_act_cohort = match_cohort_iterative(cases_dat = cases_dat,
                                                      controls_dat = controls_act,
                                                      match_rate = 5,
                                                      window_range = c(180, 180),
                                                      strict_cols = c("sex"),
                                                      flex_cols = c("age"),
                                                      fixed_baselines=c("bmi"),
                                                      range_cols = list("age"=5), 
                                                      baseline_cols = c("steps"),
                                                      baseline_range = c(-180, 0),
                                                      anchor_col_name = "date",
                                                      min_days_in_baseline = 30,
                                                      rank="bmi")
}
}
