% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_edu.R
\name{recode_edu}
\alias{recode_edu}
\title{Recode education}
\usage{
recode_edu(education_status)
}
\arguments{
\item{education_status}{character vector corresponding to the education level}
}
\description{
Recode education
}
