% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_race_wbo.R
\name{recode_race_wbo}
\alias{recode_race_wbo}
\title{Recode race to White, Black, Other}
\usage{
recode_race_wbo(race)
}
\arguments{
\item{race}{a character vector corresponding to the race column in the person table. Sets missing/skipped responses to NA.}
}
\description{
Recode race to White, Black, Other
}
