% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_sex_fm.R
\name{recode_sex_fm}
\alias{recode_sex_fm}
\title{Recode sex to Male, Female}
\usage{
recode_sex_fm(sex)
}
\arguments{
\item{sex}{character vector corresponding to the gender or sex column in the person table. Sets missing/skipped responses to NA.}
}
\description{
Recode sex to Male, Female
}
