% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_smoking.R
\name{recode_smoking}
\alias{recode_smoking}
\title{Recode smoking status}
\usage{
recode_smoking(smoking_status)
}
\arguments{
\item{smoking_status}{character vector corresponding to the smoking status}
}
\description{
Recode smoking status
}
