% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_packages.R
\name{restore_packages}
\alias{restore_packages}
\title{Restore packages to workspace}
\usage{
restore_packages()
}
\description{
Restore packages to workspace
}
\examples{
\dontrun{
backup_packages() # must be run at least once before
restore_packages()
}
}
