% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_date_window.R
\name{subset_by_date_window}
\alias{subset_by_date_window}
\title{Subset fitbit data using an anchor date and a date range}
\usage{
subset_by_date_window(
  dat,
  anchor_date,
  date_range,
  date_col = "date",
  id_col = "person_id"
)
}
\arguments{
\item{anchor_date}{the column name to be used as anchor}

\item{fitbit_dat}{fitbit data data table}

\item{before}{time range in the past}

\item{after}{time range in the future}

\item{date_column}{date column for the subset}

\item{group_by}{grouping variable for anchoring}
}
\description{
Subset fitbit data using an anchor date and a date range
}
\examples{
\dontrun{
fitbit_dat <- subset_by_date_window(fitbit_dat, "covid_date", c(100,100))
}
}
